Refactor [#183] 유틸 클래스 생성 및 패키지 분리로 인한 변경사항
