+++ morib/src/main/java/org/morib/server/api/homeView/facade/HomeViewFacade.java
@@ -27,7 +27,7 @@ import org.morib.server.domain.todo.application.FetchTodoService;
 import org.morib.server.domain.todo.infra.Todo;
 import org.morib.server.domain.user.application.service.FetchUserService;
 import org.morib.server.domain.user.infra.User;
-import org.morib.server.global.common.DataUtils;
+import org.morib.server.global.common.util.DataUtils;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.time.LocalDate;
