+++ morib/src/main/java/org/morib/server/api/modalView/controller/RelationshipController.java
@@ -4,7 +4,7 @@ import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
 import org.morib.server.api.modalView.dto.CreateRelationshipRequestDto;
 import org.morib.server.api.modalView.facade.ModalViewFacade;
-import org.morib.server.global.common.ApiResponseUtil;
+import org.morib.server.global.common.util.ApiResponseUtil;
 import org.morib.server.global.common.BaseResponse;
 import org.morib.server.global.message.SuccessMessage;
 import org.morib.server.global.userauth.CustomUserDetails;
