+++ morib/src/main/java/org/morib/server/api/settingView/controller/SettingViewController.java
@@ -3,7 +3,7 @@ package org.morib.server.api.settingView.controller;
 import lombok.RequiredArgsConstructor;
 import org.morib.server.api.settingView.dto.UpdateUserProfileRequestDto;
 import org.morib.server.api.settingView.facade.SettingViewFacade;
-import org.morib.server.global.common.ApiResponseUtil;
+import org.morib.server.global.common.util.ApiResponseUtil;
 import org.morib.server.global.common.BaseResponse;
 import org.morib.server.global.message.SuccessMessage;
 import org.morib.server.global.userauth.CustomUserDetails;
