+++ morib/src/main/java/org/morib/server/global/jwt/CustomJwtAuthenticationEntryPoint.java
@@ -4,7 +4,7 @@ import com.fasterxml.jackson.databind.ObjectMapper;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 import lombok.RequiredArgsConstructor;
-import org.morib.server.global.common.ApiResponseUtil;
+import org.morib.server.global.common.util.ApiResponseUtil;
 import org.morib.server.global.message.ErrorMessage;
 import org.springframework.http.MediaType;
 import org.springframework.security.core.AuthenticationException;
